Refactor UI and update library handling
