+++ ui/main_window.py
@@ -1,14 +1,14 @@
 from PyQt5.QtWidgets import QApplication, QWidget, QLabel, QLineEdit, QPushButton, QVBoxLayout, QMessageBox, QComboBox, QHBoxLayout
+from core.utility import User_library
+import sys
 
 class Main_window(QWidget):
     def __init__(self, parent=None):
-        super(SkipUpdateApp, self).__init__(parent)
-        
-        self.steamapps_directory = find_steamapps()
-        self.steamclient = steam.client.SteamClient()
-        self.steamclient.anonymous_login()
-        self.cdnclient = steam.client.cdn.CDNClient(self.steamclient)
-        
+        super(Main_window, self).__init__(parent)
+        self.user_library = User_library()
+        self.user_library.get_steamapps_path()
+        self.user_library.gather_library()
+
         self.initUI()
 
     def initUI(self):
@@ -27,7 +27,7 @@ class Main_window(QWidget):
             
             game_label = QLabel(self)
             skip_button = QPushButton("Skip Update", self)
-            skip_button.clicked.connect(self.make_skip_update_handler(game_label.text()))
+            # skip_button.clicked.connect(self.make_skip_update_handler(game_label.text()))
 
             game_layout.addWidget(game_label)
             game_layout.addWidget(skip_button)
@@ -41,44 +41,46 @@ class Main_window(QWidget):
 
         self.update_display()
 
-    def make_skip_update_handler(self, game_name):
-        def handler():
-            try:
-                skip_update(self.steamapps_directory, game_name, self.cdnclient)
-                QMessageBox.information(self, "Success", f"Update skipped for {game_name}!")
-                self.update_display()
-            except Exception as e:
-                QMessageBox.critical(self, "Error", str(e))
+    # def make_skip_update_handler(self, game_name):
+    #     def handler():
+    #         try:
+    #             skip_update(self.steamapps_directory, game_name, self.cdnclient)
+    #             QMessageBox.information(self, "Success", f"Update skipped for {game_name}!")
+    #             self.update_display()
+    #         except Exception as e:
+    #             QMessageBox.critical(self, "Error", str(e))
 
-        return handler
+    #     return handler
 
     def update_display(self):
         search_query = self.search_bar.text().lower()
         
-        games = list_games(self.steamapps_directory)
-        
+        apps = self.user_library.app_list.copy()
+    
         if search_query:
-            games = [game for game in games if search_query in game["name"].lower()]
+            apps = [app for app in apps if search_query in app.app_name.lower()]
 
-        for game_label, skip_button in self.game_widgets:
+        for app_label, skip_button in self.game_widgets:
             try:
-                game = games.pop(0)
-                game_name = game["name"]
-                state_flags = game["state"]
+                app = apps.pop(0)
+                app_name = app.app_name
+                # state_flags = game["state"]
                 
-                color_map = {
-                    '4': ('green', 'Has no updates'),
-                    '6': ('red', 'Has an update')
-                }
-                color, state_msg = color_map.get(state_flags, ('yellow', 'Other'))
+                # color_map = {
+                #     '4': ('green', 'Has no updates'),
+                #     '6': ('red', 'Has an update')
+                # }
+                # color, state_msg = color_map.get(state_flags, ('yellow', 'Other'))
 
-                game_label.setText(f"{game_name} - {state_msg}")
-                game_label.setStyleSheet(f"color: {color}")
+                # game_label.setText(f"{game_name} - {state_msg}")
+                app_label.setText(f"{app_name}")
+                # game_label.setStyleSheet(f"color: {color}")
                 
                 skip_button.show()
-                skip_button.clicked.disconnect()
-                skip_button.clicked.connect(self.make_skip_update_handler(game_name))
+                # skip_button.clicked.disconnect()
+                # skip_button.clicked.connect(self.make_skip_update_handler(app_name))
 
             except IndexError:
-                game_label.clear()
+                app_label.clear()
                 skip_button.hide()
+
